subscriber/requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
google-api-core==2.0.1
google-auth==2.0.2
google-cloud-core==2.0.0
google-cloud-pubsub==2.7.1
google-cloud-storage==1.42.0
google-crc32c==1.1.4
google-resumable-media==2.0.1
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.39.0
idna==3.2
libcst==0.3.20
mypy-extensions==0.4.3
mysql-connector-python==8.0.26
packaging==21.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.6
